Add Google cloudsql-proxy chart
In order for the service to communicate with the PostgreSQL database it
must authenticate with the Google Cloud SQL API and make a secure
request to the database instance. This must be done through the Google
Cloud SQL proxy, a container that manages authentication and proxies
each request made to the database through a secure channel.

This is deployed as a separate service using a Helm chart.